moving obj_data out of merc.h
